flux/ai/agent-screenshot/src/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
cloudevents>=1.10.0
structlog>=23.2.0
httpx>=0.25.0
# OCR
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
# Vision (Ollama for local, OpenAI/Anthropic optional)
openai>=1.0.0
anthropic>=0.18.0
numpy>=1.24.0
